moved each module into a separate file. Much cleaner this way
